Address -Wtautological-compare warnings

+++ src/mc-fle2-payload-iev.c
@@ -183,7 +183,7 @@ mc_FLE2IndexedEncryptedValue_write (
    const uint8_t subtype =
       (uint8_t) MC_SUBTYPE_FLE2IndexedEqualityEncryptedValue;
 
-   if ((original_bson_type < 0) || (original_bson_type > 0xFF)) {
+   if (((int) original_bson_type < 0) || ((int) original_bson_type > 0xFF)) {
       CLIENT_ERR ("Field 't' must be a valid BSON type, got: %d",
                   original_bson_type);
       CHECK_AND_GOTO (false);
